#305 - slim calypso-sim tests to a confirmable core
Replace the sprawling suite with a lean one where each assertion is
independently verifiable:

- delete scenarios.rs (a VcuMock mirroring unverifiable Cerberus pedal /
  fault constants; s3 only asserted the mock's own gate) and fold the
  protocol + ownership-isolation integration tests into one dedicated
  tests/stream.rs (harness inlined)
- trim inline unit tests: registry 5->3, keymap 10->6, cli 3->1
  (publish unchanged)

Net 29 -> 18 tests. No JSON-RPC protocol coverage lost; the dropped
cases are low-risk edges (snapshot sort order, empty-sequence /
multi-char keymap guards, unbounded increment). README Testing section
updated to match.
